+++ tests/Dummy/BasicDto.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Spatie\DataTransferObject\Tests\Dummy;
+
+use Spatie\DataTransferObject\DataTransferObject;
+
+class BasicDto extends DataTransferObject
+{
+    public string $name;
+}