+++ tests/Dummy/ComplexDto.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Spatie\DataTransferObject\Tests\Dummy;
+
+use Spatie\DataTransferObject\DataTransferObject;
+
+class ComplexDto extends DataTransferObject
+{
+    public string $name;
+
+    public BasicDto $other;
+}